+++ packages/plugin-network-capture-browser/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.4](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-network-capture-browser@1.4.3...@amplitude/plugin-network-capture-browser@1.4.4) (2025-08-08)
+
+**Note:** Version bump only for package @amplitude/plugin-network-capture-browser
+
+
+
+
+
+## [1.4.3](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-network-capture-browser@1.4.2...@amplitude/plugin-network-capture-browser@1.4.3) (2025-08-05)
+
+**Note:** Version bump only for package @amplitude/plugin-network-capture-browser
+
+
+
+
+
+## [1.4.2](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-network-capture-browser@1.4.1...@amplitude/plugin-network-capture-browser@1.4.2) (2025-07-29)
+
+**Note:** Version bump only for package @amplitude/plugin-network-capture-browser
+
+
+
+
+
 ## [1.4.1](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-network-capture-browser@1.4.0...@amplitude/plugin-network-capture-browser@1.4.1) (2025-07-17)
 
 **Note:** Version bump only for package @amplitude/plugin-network-capture-browser